torchrun\basic modes, run verbose fixes, logs less default mode,

+++ kubr/commands/rm.py
@@ -2,7 +2,7 @@ from rich import print
 
 from kubr.backends.base import JobOperationStatus
 from kubr.commands.base import BaseCommand
-from kubr.commands.utils.drawing import mascot_message
+from kubr.commands.utils.reply import confirmation_prompt, mascot_message
 
 
 class RmCommand(BaseCommand):
@@ -15,11 +15,12 @@ class RmCommand(BaseCommand):
 
     def __call__(self, job_name: str, namespace: str = "default"):
         # TODO [rm] add regular expression matching for job names
-        # TODO [rm] add confirmation prompt
         # TODO [rm] add completion for namespace typing
         try:
-            status = self.backend.delete_job(job_name=job_name, namespace=namespace)
-        except Exception:
+            if confirmation_prompt(f"Are you sure you want to delete job {job_name}?"):
+                status = self.backend.delete_job(job_name=job_name, namespace=namespace)
+        except Exception as e:
+            print(e)
             print(mascot_message(f"Job {job_name} deletion failed!"))
             return
